fc-agent: predict kernel change, increase predicted update time to 10min
Add reboots caused by kernel updates to the channel update message
(delivered via mail and displayed in the directory) and set predicted
time to 10 minutes because 5 may be too short for some services with
long startup time.

 #PL-129798
